swish old/requirement.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
appdirs==1.4.4
apple-music-python==1.0.5
asgiref==3.5.0
astroid==2.8.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
blinker==1.4
bottle==0.12.19
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
decorating==0.6.1
decorator==4.4.2
dnspython==2.2.0
Flask==2.0.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imageio==2.16.0
imageio-ffmpeg==0.4.5
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lavalink==3.1.5
lazy-object-proxy==1.6.0
mal==0.4.0
MarkupSafe==2.0.1
mccabe==0.6.1
moviepy==1.0.3
multidict==5.2.0
mutagen==1.45.1
nest-asyncio==1.5.4
numpy==1.22.2
orjson==3.6.6
Pillow==8.4.0
platformdirs==2.4.0
priority==2.0.0
proglog==0.1.9
py-cord==2.0.0a4780+g1336473e
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.13.0
PyJWT==2.3.0
pylint==2.11.1
pymongo==4.0.1
PyNaCl==1.5.0
python-dotenv==0.19.2
pytz==2021.3
quart==0.16.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
spotipy==2.19.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wavelink==1.0.2
websockets==10.1
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
xmltodict==0.12.0
xmltojson==0.2.0
yarl==1.7.0
youtube-dl==2021.6.6
yt-dlp==2022.1.21
